remove old strategy
