+++ apps/server/src/modules/legacy-school/domain/provisioning-strategy-options.ts
@@ -3,6 +3,5 @@ import { type AnyProvisioningOptions } from './school-system-options.do';
 import { SchulConneXProvisioningOptions } from './schulconnex-provisionin-options.do';
 
 export const provisioningStrategyOptions: Map<SystemProvisioningStrategy, new () => AnyProvisioningOptions> = new Map([
-	[SystemProvisioningStrategy.SCHULCONNEX_LEGACY, SchulConneXProvisioningOptions],
 	[SystemProvisioningStrategy.SCHULCONNEX_ASYNC, SchulConneXProvisioningOptions],
 ]);
