+++ apps/server/src/modules/legacy-school/testing/school-system-options-entity.factory.ts
@@ -10,7 +10,7 @@ export const schoolSystemOptionsEntityFactory = BaseFactory.define<
 >(SchoolSystemOptionsEntity, () => {
 	return {
 		school: schoolEntityFactory.buildWithId(),
-		system: systemEntityFactory.buildWithId({ provisioningStrategy: SystemProvisioningStrategy.SCHULCONNEX_LEGACY }),
+		system: systemEntityFactory.buildWithId({ provisioningStrategy: SystemProvisioningStrategy.SCHULCONNEX_ASYNC }),
 		provisioningOptions: {
 			groupProvisioningOtherEnabled: false,
 			groupProvisioningCoursesEnabled: false,
