+++ apps/server/src/modules/provisioning/strategy/schulconnex/index.ts
@@ -1,3 +1,2 @@
-export { SchulconnexSyncProvisioningStrategy } from './schulconnex-sync-provisioning.strategy';
 export { SchulconnexAsyncProvisioningStrategy } from './schulconnex-async-provisioning.strategy';
 export { SchulconnexResponseMapper } from './schulconnex-response-mapper';
