+++ apps/server/src/modules/provisioning/strategy/schulconnex/service/schulconnex-group-provisioning.service.spec.ts
@@ -18,7 +18,6 @@ import { UserService } from '@modules/user';
 import { userDoFactory } from '@modules/user/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { NotFoundLoggableException } from '@shared/common/loggable-exception';
-import { ExternalSource, Page } from '@shared/domain/domainobject';
 import { SchoolForGroupNotFoundLoggable, UserForGroupNotFoundLoggable } from '../../../loggable';
 import { SchulconnexCourseSyncService } from './schulconnex-course-sync.service';
 import { SchulconnexGroupProvisioningService } from './schulconnex-group-provisioning.service';
@@ -639,296 +638,6 @@ describe(SchulconnexGroupProvisioningService.name, () => {
 						],
 					},
 				});
-			});
-		});
-	});
-
-	describe('removeExternalGroupsAndAffiliation', () => {
-		describe('when group membership of user has not changed', () => {
-			const setup = () => {
-				const systemId = 'systemId';
-				const externalUserId = 'externalUserId';
-				const role = roleFactory.buildWithId();
-				const user = userDoFactory.buildWithId({ roles: [role], externalId: externalUserId });
-
-				const existingGroups = groupFactory.buildList(2, {
-					users: [{ userId: user.id as string, roleId: role.id }],
-				});
-
-				const firstExternalGroup = externalGroupDtoFactory.build({
-					externalId: existingGroups[0].externalSource?.externalId,
-					user: { externalUserId, roleName: role.name },
-				});
-				const secondExternalGroup = externalGroupDtoFactory.build({
-					externalId: existingGroups[1].externalSource?.externalId,
-					user: { externalUserId, roleName: role.name },
-				});
-				const externalGroups = [firstExternalGroup, secondExternalGroup];
-
-				userService.findByExternalId.mockResolvedValue(user);
-				groupService.findGroups.mockResolvedValue(new Page<Group>(existingGroups, 2));
-
-				return {
-					externalGroups,
-					systemId,
-					externalUserId,
-					user,
-				};
-			};
-
-			it('should find groups', async () => {
-				const { externalGroups, systemId, externalUserId, user } = setup();
-
-				await service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-				expect(groupService.findGroups).toHaveBeenCalledWith({ userId: user.id });
-			});
-
-			it('should not save the group', async () => {
-				const { externalGroups, systemId, externalUserId } = setup();
-
-				await service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-				expect(groupService.save).not.toHaveBeenCalled();
-			});
-
-			it('should not delete the group', async () => {
-				const { externalGroups, systemId, externalUserId } = setup();
-
-				await service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-				expect(groupService.delete).not.toHaveBeenCalled();
-			});
-		});
-
-		describe('when user is not part of a group anymore', () => {
-			describe('when group is empty after removal of the User and is not synced to a course', () => {
-				const setup = () => {
-					const systemId = 'systemId';
-					const externalUserId = 'externalUserId';
-					const role = roleFactory.buildWithId();
-					const user = userDoFactory.buildWithId({ roles: [role], externalId: externalUserId });
-					const firstExistingGroup = groupFactory.build({
-						users: [{ userId: user.id as string, roleId: role.id }],
-						externalSource: new ExternalSource({
-							externalId: 'externalId-1',
-							systemId,
-							lastSyncedAt: mockDate,
-						}),
-					});
-					const secondExistingGroup = groupFactory.build({
-						users: [{ userId: user.id as string, roleId: role.id }],
-						externalSource: new ExternalSource({
-							externalId: 'externalId-2',
-							systemId,
-							lastSyncedAt: mockDate,
-						}),
-					});
-					const existingGroups = [firstExistingGroup, secondExistingGroup];
-
-					const firstExternalGroup = externalGroupDtoFactory.build({
-						externalId: existingGroups[0].externalSource?.externalId,
-						user: { externalUserId, roleName: role.name },
-					});
-					const externalGroups = [firstExternalGroup];
-
-					userService.findByExternalId.mockResolvedValue(user);
-					groupService.findGroups.mockResolvedValue(new Page<Group>(existingGroups, 2));
-					courseService.findBySyncedGroup.mockResolvedValue([]);
-					return {
-						externalGroups,
-						systemId,
-						externalUserId,
-						existingGroups,
-					};
-				};
-
-				it('should delete the group', async () => {
-					const { externalGroups, systemId, externalUserId, existingGroups } = setup();
-
-					await service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-					expect(groupService.delete).toHaveBeenCalledWith(existingGroups[1]);
-				});
-
-				it('should not save the group', async () => {
-					const { externalGroups, systemId, externalUserId } = setup();
-
-					await service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-					expect(groupService.save).not.toHaveBeenCalled();
-				});
-
-				it('should not return a group', async () => {
-					const { externalGroups, systemId, externalUserId } = setup();
-
-					const result = await service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-					expect(result).toHaveLength(0);
-				});
-			});
-
-			describe('when group is empty after removal of the User but synced with a course', () => {
-				const setup = () => {
-					const systemId = new ObjectId().toHexString();
-					const externalUserId = 'externalUserId';
-					const role = roleFactory.buildWithId();
-					const user = userDoFactory.buildWithId({ roles: [role], externalId: externalUserId });
-					const course = courseFactory.build();
-
-					const externalGroupId = 'external-group-id';
-
-					const group = groupFactory.build({
-						users: [{ userId: user.id as string, roleId: role.id }],
-						externalSource: new ExternalSource({
-							systemId,
-							externalId: externalGroupId,
-							lastSyncedAt: new Date(),
-						}),
-					});
-
-					const externalGroup = externalGroupDtoFactory.build({
-						user: { externalUserId, roleName: role.name },
-						externalId: `other-${externalUserId}`,
-					});
-
-					userService.findByExternalId.mockResolvedValueOnce(user);
-					groupService.findGroups.mockResolvedValueOnce(new Page<Group>([group], 1));
-					courseService.findBySyncedGroup.mockResolvedValueOnce([course]);
-
-					return {
-						externalUserId,
-						externalGroup,
-						systemId,
-						group,
-						course,
-					};
-				};
-
-				it('should desync the course and create a history of the sync', async () => {
-					const { externalUserId, externalGroup, systemId, group, course } = setup();
-
-					await service.removeExternalGroupsAndAffiliation(externalUserId, [externalGroup], systemId);
-
-					expect(schulconnexCourseSyncService.desyncCoursesAndCreateHistories).toBeCalledTimes(1);
-					expect(schulconnexCourseSyncService.desyncCoursesAndCreateHistories).toBeCalledWith(group, [course]);
-				});
-
-				it('should delete the empty group', async () => {
-					const { externalUserId, externalGroup, systemId, group } = setup();
-
-					await service.removeExternalGroupsAndAffiliation(externalUserId, [externalGroup], systemId);
-
-					expect(groupService.delete).toBeCalledTimes(1);
-					expect(groupService.delete).toBeCalledWith(group);
-				});
-			});
-
-			describe('when group is not empty after removal of the User', () => {
-				const setup = () => {
-					const systemId = 'systemId';
-					const externalUserId = 'externalUserId';
-					const anotherExternalUserId = 'anotherExternalUserId';
-					const role = roleFactory.buildWithId();
-					const user = userDoFactory.buildWithId({ roles: [role], externalId: externalUserId });
-					const anotherUser = userDoFactory.buildWithId({ roles: [role], externalId: anotherExternalUserId });
-
-					const firstExistingGroup = groupFactory.build({
-						users: [
-							{ userId: user.id as string, roleId: role.id },
-							{ userId: anotherUser.id as string, roleId: role.id },
-						],
-						externalSource: new ExternalSource({
-							externalId: `externalId-1`,
-							systemId,
-							lastSyncedAt: mockDate,
-						}),
-					});
-
-					const secondExistingGroup = groupFactory.build({
-						users: [
-							{ userId: user.id as string, roleId: role.id },
-							{ userId: anotherUser.id as string, roleId: role.id },
-						],
-						externalSource: new ExternalSource({
-							externalId: `externalId-2`,
-							systemId,
-							lastSyncedAt: mockDate,
-						}),
-					});
-
-					const existingGroups = [firstExistingGroup, secondExistingGroup];
-
-					const firstExternalGroup = externalGroupDtoFactory.build({
-						externalId: existingGroups[0].externalSource?.externalId,
-						user: { externalUserId, roleName: role.name },
-					});
-					const externalGroups = [firstExternalGroup];
-
-					userService.findByExternalId.mockResolvedValueOnce(user);
-					groupService.findGroups.mockResolvedValueOnce(new Page<Group>(existingGroups, 2));
-					groupService.save.mockResolvedValueOnce(secondExistingGroup);
-
-					return {
-						externalGroups,
-						systemId,
-						externalUserId,
-						existingGroups,
-						secondExistingGroup,
-					};
-				};
-
-				it('should save the group', async () => {
-					const { externalGroups, systemId, externalUserId, existingGroups } = setup();
-
-					await service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-					expect(groupService.save).toHaveBeenCalledWith(existingGroups[1]);
-				});
-
-				it('should not delete the group', async () => {
-					const { externalGroups, systemId, externalUserId } = setup();
-
-					await service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-					expect(groupService.delete).not.toHaveBeenCalled();
-				});
-
-				it('should return the modified groups', async () => {
-					const { externalGroups, systemId, externalUserId, secondExistingGroup } = setup();
-
-					const result: Group[] = await service.removeExternalGroupsAndAffiliation(
-						externalUserId,
-						externalGroups,
-						systemId
-					);
-
-					expect(result).toEqual([secondExistingGroup]);
-				});
-			});
-		});
-
-		describe('when user could not be found', () => {
-			const setup = () => {
-				const systemId = 'systemId';
-				const externalUserId = 'externalUserId';
-				const externalGroups = [externalGroupDtoFactory.build()];
-
-				userService.findByExternalId.mockResolvedValue(null);
-
-				return {
-					systemId,
-					externalUserId,
-					externalGroups,
-				};
-			};
-
-			it('should throw NotFoundLoggableException', async () => {
-				const { externalGroups, systemId, externalUserId } = setup();
-
-				const func = () => service.removeExternalGroupsAndAffiliation(externalUserId, externalGroups, systemId);
-
-				await expect(func).rejects.toThrow(new NotFoundLoggableException('User', { externalId: externalUserId }));
 			});
 		});
 	});
