+++ apps/server/src/shared/domain/interface/system-provisioning.strategy.ts
@@ -1,5 +1,4 @@
 export enum SystemProvisioningStrategy {
-	SCHULCONNEX_LEGACY = 'sanis',
 	SCHULCONNEX_ASYNC = 'schulconnex-async',
 	OIDC = 'oidc',
 	TSP = 'tsp',
